move user-related calls from EthService to its own class
