+++ src/main/scala/io/iohk/ethereum/jsonrpc/EthJsonMethodsImplicits.scala
@@ -217,20 +217,6 @@ object EthJsonMethodsImplicits extends JsonMethodsImplicits {
 
   }
 
-  implicit val eth_getCode = new JsonMethodDecoder[GetCodeRequest] with JsonEncoder[GetCodeResponse] {
-    def decodeJson(params: Option[JArray]): Either[JsonRpcError, GetCodeRequest] =
-      params match {
-        case Some(JArray((address: JString) :: (blockValue: JValue) :: Nil)) =>
-          for {
-            addr <- extractAddress(address)
-            block <- extractBlockParam(blockValue)
-          } yield GetCodeRequest(addr, block)
-        case _ => Left(InvalidParams())
-      }
-
-    def encodeJson(t: GetCodeResponse): JValue = encodeAsHex(t.result)
-  }
-
   implicit val eth_getUncleCountByBlockNumber = new JsonMethodDecoder[GetUncleCountByBlockNumberRequest]
     with JsonEncoder[GetUncleCountByBlockNumberResponse] {
     def decodeJson(params: Option[JArray]): Either[JsonRpcError, GetUncleCountByBlockNumberRequest] =
@@ -271,51 +257,6 @@ object EthJsonMethodsImplicits extends JsonMethodsImplicits {
       }
 
     def encodeJson(t: GetBlockTransactionCountByNumberResponse): JValue = encodeAsHex(t.result)
-  }
-
-  implicit val eth_getBalance = new JsonMethodDecoder[GetBalanceRequest] with JsonEncoder[GetBalanceResponse] {
-    def decodeJson(params: Option[JArray]): Either[JsonRpcError, GetBalanceRequest] =
-      params match {
-        case Some(JArray((addressStr: JString) :: (blockValue: JValue) :: Nil)) =>
-          for {
-            address <- extractAddress(addressStr)
-            block <- extractBlockParam(blockValue)
-          } yield GetBalanceRequest(address, block)
-        case other =>
-          Left(InvalidParams())
-      }
-
-    def encodeJson(t: GetBalanceResponse): JValue = encodeAsHex(t.value)
-  }
-
-  implicit val eth_getStorageAt = new JsonMethodDecoder[GetStorageAtRequest] with JsonEncoder[GetStorageAtResponse] {
-    def decodeJson(params: Option[JArray]): Either[JsonRpcError, GetStorageAtRequest] =
-      params match {
-        case Some(JArray((addressStr: JString) :: (positionStr: JString) :: (blockValue: JValue) :: Nil)) =>
-          for {
-            address <- extractAddress(addressStr)
-            position <- extractQuantity(positionStr)
-            block <- extractBlockParam(blockValue)
-          } yield GetStorageAtRequest(address, position, block)
-        case _ => Left(InvalidParams())
-      }
-
-    def encodeJson(t: GetStorageAtResponse): JValue = encodeAsHex(t.value)
-  }
-
-  implicit val eth_getTransactionCount = new JsonMethodDecoder[GetTransactionCountRequest]
-    with JsonEncoder[GetTransactionCountResponse] {
-    def decodeJson(params: Option[JArray]): Either[JsonRpcError, GetTransactionCountRequest] =
-      params match {
-        case Some(JArray((addressStr: JString) :: (blockValue: JValue) :: Nil)) =>
-          for {
-            address <- extractAddress(addressStr)
-            block <- extractBlockParam(blockValue)
-          } yield GetTransactionCountRequest(address, block)
-        case _ => Left(InvalidParams())
-      }
-
-    def encodeJson(t: GetTransactionCountResponse): JValue = encodeAsHex(t.value)
   }
 
   implicit val newFilterResponseEnc = new JsonEncoder[NewFilterResponse] {
@@ -485,35 +426,6 @@ object EthJsonMethodsImplicits extends JsonMethodsImplicits {
       value = value.getOrElse(0),
       data = data.getOrElse(ByteString(""))
     )
-  }
-
-  implicit val eth_getStorageRoot = new JsonMethodDecoder[GetStorageRootRequest]
-    with JsonEncoder[GetStorageRootResponse] {
-    def decodeJson(params: Option[JArray]): Either[JsonRpcError, GetStorageRootRequest] =
-      params match {
-        case Some(JArray((addressStr: JString) :: (blockValue: JValue) :: Nil)) =>
-          for {
-            address <- extractAddress(addressStr)
-            block <- extractBlockParam(blockValue)
-          } yield GetStorageRootRequest(address, block)
-        case _ => Left(InvalidParams())
-      }
-
-    def encodeJson(t: GetStorageRootResponse): JValue = encodeAsHex(t.storageRoot)
-  }
-
-  def extractStorageKeys(input: JValue): Either[JsonRpcError, Seq[StorageProofKey]] = {
-    import cats.syntax.traverse._
-    import cats.syntax.either._
-    input match {
-      case JArray(elems) =>
-        elems.traverse { x =>
-          extractQuantity(x)
-            .map(StorageProofKey.apply)
-            .leftMap(_ => InvalidParams(s"Invalid param storage proof key: $x"))
-        }
-      case _ => Left(InvalidParams())
-    }
   }
 
   implicit val eth_getProof: JsonMethodDecoder[GetProofRequest] with JsonEncoder[GetProofResponse] =
