+++ src/main/scala/io/iohk/ethereum/jsonrpc/EthService.scala
@@ -85,20 +85,6 @@ object EthService {
   case class IeleCallResponse(returnData: Seq[ByteString])
   case class EstimateGasResponse(gas: BigInt)
 
-  case class GetCodeRequest(address: Address, block: BlockParam)
-  case class GetCodeResponse(result: ByteString)
-
-  case class GetBalanceRequest(address: Address, block: BlockParam)
-  case class GetBalanceResponse(value: BigInt)
-
-  case class GetStorageAtRequest(address: Address, position: BigInt, block: BlockParam)
-  case class GetStorageAtResponse(value: ByteString)
-
-  case class ResolvedBlock(block: Block, pendingState: Option[InMemoryWorldStateProxy])
-
-  case class GetTransactionCountRequest(address: Address, block: BlockParam)
-  case class GetTransactionCountResponse(value: BigInt)
-
   case class NewFilterRequest(filter: Filter)
   case class Filter(
       fromBlock: Option[BlockParam],
@@ -124,8 +110,31 @@ object EthService {
   case class GetLogsRequest(filter: Filter)
   case class GetLogsResponse(filterLogs: LogFilterLogs)
 
-  case class GetStorageRootRequest(address: Address, block: BlockParam)
-  case class GetStorageRootResponse(storageRoot: ByteString)
+  case class ResolvedBlock(block: Block, pendingState: Option[InMemoryWorldStateProxy])
+
+  def resolveBlock(
+      blockchain: Blockchain,
+      ledger: Ledger,
+      blockParam: BlockParam
+  ): Either[JsonRpcError, ResolvedBlock] = {
+    def getBlock(number: BigInt): Either[JsonRpcError, Block] = {
+      blockchain
+        .getBlockByNumber(number)
+        .map(Right.apply)
+        .getOrElse(Left(JsonRpcError.InvalidParams(s"Block $number not found")))
+    }
+
+    blockParam match {
+      case BlockParam.WithNumber(blockNumber) => getBlock(blockNumber).map(ResolvedBlock(_, pendingState = None))
+      case BlockParam.Earliest => getBlock(0).map(ResolvedBlock(_, pendingState = None))
+      case BlockParam.Latest => getBlock(blockchain.getBestBlockNumber()).map(ResolvedBlock(_, pendingState = None))
+      case BlockParam.Pending =>
+        ledger.consensus.blockGenerator.getPendingBlockAndState
+          .map(pb => ResolvedBlock(pb.pendingBlock.block, pendingState = Some(pb.worldState)))
+          .map(Right.apply)
+          .getOrElse(resolveBlock(blockchain, ledger, BlockParam.Latest)) //Default behavior in other clients
+    }
+  }
 
   def resolveBlock(
       blockParam: BlockParam,
@@ -240,38 +249,6 @@ class EthService(
     }
   }
 
-  def getCode(req: GetCodeRequest): ServiceResponse[GetCodeResponse] = {
-    Task {
-      resolveBlock(req.block, blockchain, blockGenerator).map { case ResolvedBlock(block, _) =>
-        val world = blockchain.getWorldStateProxy(
-          block.header.number,
-          blockchainConfig.accountStartNonce,
-          block.header.stateRoot,
-          noEmptyAccounts = false,
-          ethCompatibleStorage = blockchainConfig.ethCompatibleStorage
-        )
-        GetCodeResponse(world.getCode(req.address))
-      }
-    }
-  }
-
-  def getBalance(req: GetBalanceRequest): ServiceResponse[GetBalanceResponse] =
-    withAccount(req.address, req.block) { account =>
-      GetBalanceResponse(account.balance)
-    }
-
-  def getStorageAt(req: GetStorageAtRequest): ServiceResponse[GetStorageAtResponse] =
-    withAccount(req.address, req.block) { account =>
-      GetStorageAtResponse(
-        blockchain.getAccountStorageAt(account.storageRoot, req.position, blockchainConfig.ethCompatibleStorage)
-      )
-    }
-
-  def getTransactionCount(req: GetTransactionCountRequest): ServiceResponse[GetTransactionCountResponse] =
-    withAccount(req.address, req.block) { account =>
-      GetTransactionCountResponse(account.nonce)
-    }
-
   def newFilter(req: NewFilterRequest): ServiceResponse[NewFilterResponse] = {
     implicit val timeout: Timeout = Timeout(filterConfig.filterManagerQueryTimeout)
 
@@ -339,29 +316,16 @@ class EthService(
       }
   }
 
-  private def withAccount[T](address: Address, blockParam: BlockParam)(makeResponse: Account => T): ServiceResponse[T] =
-    Task {
-      resolveBlock(blockParam, blockchain, blockGenerator)
-        .map { case ResolvedBlock(block, _) =>
-          blockchain
-            .getAccount(address, block.header.number)
-            .getOrElse(Account.empty(blockchainConfig.accountStartNonce))
-        }
-        .map(makeResponse)
-    }.onErrorRecover { case _: MissingNodeException =>
-      Left(JsonRpcError.NodeNotFound)
-    }
-
   private def doCall[A](req: CallRequest)(
       f: (SignedTransactionWithSender, BlockHeader, Option[InMemoryWorldStateProxy]) => A
   ): Either[JsonRpcError, A] = for {
     stx <- prepareTransaction(req)
-    block <- resolveBlock(req.block, blockchain, blockGenerator)
+    block <- resolveBlock(blockchain, ledger, req.block)
   } yield f(stx, block.block.header, block.pendingState)
 
   private def getGasLimit(req: CallRequest): Either[JsonRpcError, BigInt] =
     if (req.tx.gas.isDefined) Right[JsonRpcError, BigInt](req.tx.gas.get)
-    else resolveBlock(BlockParam.Latest, blockchain, blockGenerator).map(r => r.block.header.gasLimit)
+    else resolveBlock(blockchain, ledger, BlockParam.Latest).map(r => r.block.header.gasLimit)
 
   private def prepareTransaction(req: CallRequest): Either[JsonRpcError, SignedTransactionWithSender] = {
     getGasLimit(req).map { gasLimit =>
@@ -382,10 +346,5 @@ class EthService(
       SignedTransactionWithSender(tx, fakeSignature, fromAddress)
     }
   }
-
-  def getStorageRoot(req: GetStorageRootRequest): ServiceResponse[GetStorageRootResponse] =
-    withAccount(req.address, req.block) { account =>
-      GetStorageRootResponse(account.storageRoot)
-    }
 
 }
