+++ src/main/scala/io/iohk/ethereum/nodebuilder/NodeBuilder.scala
@@ -423,16 +423,8 @@ trait EthMiningServiceBuilder {
 trait EthTxServiceBuilder {
   self: StorageBuilder
     with BlockchainBuilder
-    // with BlockchainConfigBuilder
     with PendingTransactionsManagerBuilder
     with LedgerBuilder
-    // with KeyStoreBuilder
-    // with SyncControllerBuilder
-    // with OmmersPoolBuilder
-    // with ConsensusBuilder
-    // with ConsensusConfigBuilder
-    // with FilterManagerBuilder
-    // with FilterConfigBuilder
     with TxPoolConfigBuilder =>
 
   lazy val ethTxService = new EthTxService(
@@ -447,6 +439,16 @@ trait EthBlocksServiceBuilder {
   self: BlockchainBuilder with BlockchainConfigBuilder with LedgerBuilder =>
 
   lazy val ethBlocksService = new EthBlocksService(blockchain, ledger, blockchainConfig)
+}
+
+trait EthUserServiceBuilder {
+  self: StorageBuilder with BlockchainBuilder with BlockchainConfigBuilder with LedgerBuilder =>
+
+  lazy val ethUserService = new EthUserService(
+    blockchain,
+    ledger,
+    blockchainConfig
+  )
 }
 
 trait PersonalServiceBuilder {
@@ -534,6 +536,7 @@ trait JSONRpcControllerBuilder {
     with EthMiningServiceBuilder
     with EthBlocksServiceBuilder
     with EthTxServiceBuilder
+    with EthUserServiceBuilder
     with NetServiceBuilder
     with PersonalServiceBuilder
     with DebugServiceBuilder
@@ -554,6 +557,7 @@ trait JSONRpcControllerBuilder {
       ethMiningService,
       ethBlocksService,
       ethTxService,
+      ethUserService,
       personalService,
       testService,
       debugService,
@@ -736,6 +740,7 @@ trait Node
     with EthMiningServiceBuilder
     with EthBlocksServiceBuilder
     with EthTxServiceBuilder
+    with EthUserServiceBuilder
     with NetServiceBuilder
     with PersonalServiceBuilder
     with DebugServiceBuilder
