+++ src/test/scala/io/iohk/ethereum/jsonrpc/CheckpointingJRCSpec.scala
@@ -211,6 +211,7 @@ class CheckpointingJRCSpec
     val ethMiningService = mock[EthMiningService]
     val ethBlocksService = mock[EthBlocksService]
     val ethTxService = mock[EthTxService]
+    val ethUserService = mock[EthUserService]
     val qaService = mock[QAService]
     val checkpointingService = mock[CheckpointingService]
     val mantisService = mock[MantisService]
@@ -223,6 +224,7 @@ class CheckpointingJRCSpec
         ethMiningService,
         ethBlocksService,
         ethTxService,
+        ethUserService,
         personalService,
         None,
         debugService,
