+++ src/test/scala/io/iohk/ethereum/jsonrpc/EthServiceSpec.scala
@@ -148,128 +148,6 @@ class EthServiceSpec
     response.runSyncUnsafe() shouldEqual Right(EstimateGasResponse(123))
   }
 
-  it should "handle getCode request" in new TestSetup {
-    (() => ledger.consensus).expects().returns(consensus)
-    val address = Address(ByteString(Hex.decode("abbb6bebfa05aa13e908eaa492bd7a8343760477")))
-    storagesInstance.storages.evmCodeStorage.put(ByteString("code hash"), ByteString("code code code")).commit()
-
-    import MerklePatriciaTrie.defaultByteArraySerializable
-
-    val mpt =
-      MerklePatriciaTrie[Array[Byte], Account](storagesInstance.storages.stateStorage.getBackingStorage(0))
-        .put(
-          crypto.kec256(address.bytes.toArray[Byte]),
-          Account(0, UInt256(0), ByteString(""), ByteString("code hash"))
-        )
-
-    val newBlockHeader = blockToRequest.header.copy(stateRoot = ByteString(mpt.getRootHash))
-    val newblock = blockToRequest.copy(header = newBlockHeader)
-    blockchain.storeBlock(newblock).commit()
-    blockchain.saveBestKnownBlocks(newblock.header.number)
-
-    val response = ethService.getCode(GetCodeRequest(address, BlockParam.Latest))
-
-    response.runSyncUnsafe() shouldEqual Right(GetCodeResponse(ByteString("code code code")))
-  }
-
-  it should "handle getBalance request" in new TestSetup {
-    (() => ledger.consensus).expects().returns(consensus)
-    val address = Address(ByteString(Hex.decode("abbb6bebfa05aa13e908eaa492bd7a8343760477")))
-
-    import MerklePatriciaTrie.defaultByteArraySerializable
-
-    val mpt =
-      MerklePatriciaTrie[Array[Byte], Account](storagesInstance.storages.stateStorage.getBackingStorage(0))
-        .put(
-          crypto.kec256(address.bytes.toArray[Byte]),
-          Account(0, UInt256(123), ByteString(""), ByteString("code hash"))
-        )
-
-    val newBlockHeader = blockToRequest.header.copy(stateRoot = ByteString(mpt.getRootHash))
-    val newblock = blockToRequest.copy(header = newBlockHeader)
-    blockchain.storeBlock(newblock).commit()
-    blockchain.saveBestKnownBlocks(newblock.header.number)
-
-    val response = ethService.getBalance(GetBalanceRequest(address, BlockParam.Latest))
-
-    response.runSyncUnsafe() shouldEqual Right(GetBalanceResponse(123))
-  }
-
-  it should "handle MissingNodeException when getting balance" in new TestSetup {
-    (() => ledger.consensus).expects().returns(consensus)
-    val address = Address(ByteString(Hex.decode("abbb6bebfa05aa13e908eaa492bd7a8343760477")))
-
-    val newBlockHeader = blockToRequest.header
-    val newblock = blockToRequest.copy(header = newBlockHeader)
-    blockchain.storeBlock(newblock).commit()
-    blockchain.saveBestKnownBlocks(newblock.header.number)
-
-    val response = ethService.getBalance(GetBalanceRequest(address, BlockParam.Latest))
-
-    response.runSyncUnsafe() shouldEqual Left(JsonRpcError.NodeNotFound)
-  }
-
-  it should "handle getStorageAt request" in new TestSetup {
-    import io.iohk.ethereum.rlp.UInt256RLPImplicits._
-
-    (() => ledger.consensus).expects().returns(consensus)
-    val address = Address(ByteString(Hex.decode("abbb6bebfa05aa13e908eaa492bd7a8343760477")))
-
-    import MerklePatriciaTrie.defaultByteArraySerializable
-
-    val byteArrayUInt256Serializer = new ByteArrayEncoder[UInt256] {
-      override def toBytes(input: UInt256): Array[Byte] = input.bytes.toArray[Byte]
-    }
-
-    val rlpUInt256Serializer = new ByteArraySerializable[UInt256] {
-      override def fromBytes(bytes: Array[Byte]): UInt256 = ByteString(bytes).toUInt256
-      override def toBytes(input: UInt256): Array[Byte] = input.toBytes
-    }
-
-    val storageMpt =
-      io.iohk.ethereum.domain.EthereumUInt256Mpt
-        .storageMpt(
-          ByteString(MerklePatriciaTrie.EmptyRootHash),
-          storagesInstance.storages.stateStorage.getBackingStorage(0)
-        )
-        .put(UInt256(333), UInt256(123))
-
-    val mpt =
-      MerklePatriciaTrie[Array[Byte], Account](storagesInstance.storages.stateStorage.getBackingStorage(0))
-        .put(
-          crypto.kec256(address.bytes.toArray[Byte]),
-          Account(0, UInt256(0), ByteString(storageMpt.getRootHash), ByteString(""))
-        )
-
-    val newBlockHeader = blockToRequest.header.copy(stateRoot = ByteString(mpt.getRootHash))
-    val newblock = blockToRequest.copy(header = newBlockHeader)
-    blockchain.storeBlock(newblock).commit()
-    blockchain.saveBestKnownBlocks(newblock.header.number)
-
-    val response = ethService.getStorageAt(GetStorageAtRequest(address, 333, BlockParam.Latest))
-    response.runSyncUnsafe().map(v => UInt256(v.value)) shouldEqual Right(UInt256(123))
-  }
-
-  it should "handle get transaction count request" in new TestSetup {
-    (() => ledger.consensus).expects().returns(consensus)
-    val address = Address(ByteString(Hex.decode("abbb6bebfa05aa13e908eaa492bd7a8343760477")))
-
-    import MerklePatriciaTrie.defaultByteArraySerializable
-
-    val mpt =
-      MerklePatriciaTrie[Array[Byte], Account](storagesInstance.storages.stateStorage.getBackingStorage(0))
-        .put(crypto.kec256(address.bytes.toArray[Byte]), Account(999, UInt256(0), ByteString(""), ByteString("")))
-
-    val newBlockHeader = blockToRequest.header.copy(stateRoot = ByteString(mpt.getRootHash))
-    val newblock = blockToRequest.copy(header = newBlockHeader)
-    blockchain.storeBlock(newblock).commit()
-    blockchain.saveBestKnownBlocks(newblock.header.number)
-
-    val response = ethService.getTransactionCount(GetTransactionCountRequest(address, BlockParam.Latest))
-
-    response.runSyncUnsafe() shouldEqual Right(GetTransactionCountResponse(BigInt(999)))
-  }
-
   // NOTE TestSetup uses Ethash consensus; check `consensusConfig`.
   class TestSetup(implicit system: ActorSystem) extends MockFactory with EphemBlockchainTestSetup with ApisBuilder {
     val blockGenerator = mock[EthashBlockGenerator]
