+++ src/test/scala/io/iohk/ethereum/jsonrpc/EthTxServiceSpec.scala
@@ -2,34 +2,34 @@ package io.iohk.ethereum.jsonrpc
 
 import akka.actor.ActorSystem
 import akka.testkit.TestKit
+import akka.testkit.TestProbe
+import akka.util.ByteString
+import io.iohk.ethereum._
+import io.iohk.ethereum.blockchain.sync.EphemBlockchainTestSetup
+import io.iohk.ethereum.consensus._
+import io.iohk.ethereum.consensus.ethash.blocks.EthashBlockGenerator
+import io.iohk.ethereum.crypto.ECDSASignature
+import io.iohk.ethereum.db.storage.AppStateStorage
+import io.iohk.ethereum.domain._
+import io.iohk.ethereum.jsonrpc.EthService.BlockParam
+import io.iohk.ethereum.jsonrpc.EthTxService._
+import io.iohk.ethereum.keystore.KeyStore
+import io.iohk.ethereum.ledger.Ledger
+import io.iohk.ethereum.nodebuilder.ApisBuilder
 import io.iohk.ethereum.NormalPatience
+import io.iohk.ethereum.transactions.PendingTransactionsManager
+import io.iohk.ethereum.transactions.PendingTransactionsManager._
+import io.iohk.ethereum.utils._
 import io.iohk.ethereum.WithActorSystemShutDown
+import monix.execution.Scheduler.Implicits.global
 import org.scalactic.TypeCheckedTripleEquals
 import org.scalamock.scalatest.MockFactory
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.flatspec.AnyFlatSpecLike
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.OptionValues
-import io.iohk.ethereum.jsonrpc.EthTxService._
 import scala.concurrent.duration.Duration
 import scala.concurrent.duration.DurationInt
-import io.iohk.ethereum.blockchain.sync.EphemBlockchainTestSetup
-import io.iohk.ethereum.nodebuilder.ApisBuilder
-import io.iohk.ethereum.db.storage.AppStateStorage
-import io.iohk.ethereum.keystore.KeyStore
-import io.iohk.ethereum.ledger.Ledger
-import io.iohk.ethereum.utils._
-import monix.execution.Scheduler.Implicits.global
-import io.iohk.ethereum.domain._
-import io.iohk.ethereum._
-import io.iohk.ethereum.jsonrpc.EthService.BlockParam
-import akka.util.ByteString
-import io.iohk.ethereum.crypto.ECDSASignature
-import io.iohk.ethereum.consensus._
-import io.iohk.ethereum.consensus.ethash.blocks.EthashBlockGenerator
-import akka.testkit.TestProbe
-import io.iohk.ethereum.transactions.PendingTransactionsManager
-import io.iohk.ethereum.transactions.PendingTransactionsManager._
 import scala.concurrent.duration.FiniteDuration
 
 class EthTxServiceSpec
